feat: extract campaign audience selection logic into dedicated hook and helper files
